BASWSPRT-112: change class name
