+++ CRM/Civicase/Upgrader.php
@@ -6,7 +6,7 @@ use CRM_Civicase_Setup_AddCaseTypesForCustomGroupExtends as AddCaseTypesForCusto
 use CRM_Civicase_Setup_AddCaseCategoryWordReplacementOptionGroup as AddCaseCategoryWordReplacementOptionGroup;
 use CRM_Civicase_Setup_MoveCaseTypesToCasesCategory as MoveCaseTypesToCasesCategory;
 use CRM_Civicase_Helper_CaseCategory as CaseCategoryHelper;
-use CRM_Civicase_Setup_CreateSafeFileExtentionOptionValue as CreateSafeFileExtentionOptionValue;
+use CRM_Civicase_Setup_CreateSafeFileExtensionOptionValue as CreateSafeFileExtensionOptionValue;
 
 /**
  * Collection of upgrade steps.
@@ -71,7 +71,7 @@ class CRM_Civicase_Upgrader extends CRM_Civicase_Upgrader_Base {
       new CreateCasesOptionValue(),
       new AddCaseTypesForCustomGroupExtends(),
       new MoveCaseTypesToCasesCategory(),
-      new CreateSafeFileExtentionOptionValue(),
+      new CreateSafeFileExtensionOptionValue(),
     ];
     foreach ($steps as $step) {
       $step->apply();
@@ -556,7 +556,7 @@ class CRM_Civicase_Upgrader extends CRM_Civicase_Upgrader_Base {
    *   The queue requires that true is returned on successful upgrade, but we
    *   use exceptions to indicate an error instead.
    */
-  public static function runStepUpgrade(CRM_Queue_TaskContext $context, object $step) {
+  public static function runStepUpgrade(CRM_Queue_TaskContext $context, $step) {
     $step->apply();
     return TRUE;
   }
